cqn4sql: localized unmanaged path expression in model
also consider the localized version of an association if the first step
of a ref is checked wheter it equals the association of the path expression

e.g. `SELECT from bookshop.Authors:books` if the `books` assoc has the on
condition: `books.<…> = ID` we need to consider `localized.bookshop.Authors:books`
to identify the first ref step in the on condition as the association of the original
path.