cds.infer: better error handling for non-fk traversal in infix filter (#40)
* cds.infer: better error handling for non-fk traversal in infix filter

in an infix filter path step contains an association, it must not be an unmanaged
association. Moreover, the next step must be a foreign key access. If not, we must
emit an error. We do so already for infix filters in `from`, but this was missing
for regular columns.

This changes also contains refactoring of the `inferQueryElement` function to
include a `context` object, containing information about the current scope of the ref.

* ✨prettier✨

* less args

* one more test