cqn4sql / infer: better alias handling (#47)
If we encounter a subquery in a column, or have to build a correlated subquery by ourselves for an `expand` column, it is important to provide the aliases of all the outer queries to the to-be-transformed subquery. This way, we can ensure table alias uniqueness across the query. This change improves this by taking the `query.outerQueries` into account when calculating the table aliases.

This mechanism only works for automated table alias calculation, if a user provides a table alias manually, `cqn4sql` does respect the alias even though it might clash with automatically created table aliases of the outer scope.