cqn4sql: replace implicit aliases in subqueries if necessary (#55)
In a subquery, there might be a need to replace any usage of an implicit alias with the actual auto-generated table alias, which is provided by cqn4sql.

This change makes sure, that the correct alias for a given `ref` with length > 1 is found if:
- the original query has **NO** explicit alias
- `ref[0]` equals the implicit alias of the query (i.e. `from.ref[ from.length - 1 ].split('.').pop()`)
- but differs from the explicit alias, assigned by `cqn4sql` (i.e. `<subquery>.from.uniqueSubqueryAlias`)

#### other, cosmetic changes

Also this change improves the calculation of the alias of a correlated expand subquery, please look at the changed tests how that affects the current behavior. Basically, when an expand is written with an explicit table alias before the association `… from bookshop.Books { Books.author { name }, …` we must not consider the table alias `Books` for the alias of the correlated subquery.