fix(draft): deep update on a managed draft entity without change should update the `modifiedAt` field on the parent entity
```cds
entity Shipment : managed, cuid {
    shipmentDate : Composition of many ShipmentDate on shipmentDate.owner = $self
}

entity ShipmentDate : cuid, managed {
    owner         : Association to Shipment;
    dateID        : Association to ShipmentDateDesc;
    dateTimeStamp : DateTime
}

entity ShipmentDateDesc {
    key dateID      : Integer;
        description : localized String(240);
}
```

```cds
namespace cgt.filter.pocservice;

using { cgt.filter.poc as shipment } from '../db/schema';

service ManageShipmentService {
  annotate Shipment with  @odata.draft.enabled  @Common.SemanticKey: [ID];
  entity Shipment as projection on shipment.Shipment;

  entity ShipmentDate as projection on shipment.ShipmentDate {
    ID,
    owner : redirected to Shipment,
    dateID : redirected to ShipmentDateDesc,
    dateTimeStamp as timestamp,
    modifiedAt
  };

  entity ShipmentDateDesc as projection on shipment.ShipmentDateDesc;
}
```

In a draft scenario, after editing a draft-enabled entity with child entities and activating the draft without any
modifications, the `modifiedAt` field of the parent entity (`ShipmentDate`) should be updated.
This means that when the user opens the Fiori Element-based application (List Report Floorplan), clicks on a list item
to navigate to the Fiori Object Page Floorplan, and then presses the "Edit" button to edit the draft, if no changes
are made to the child entities, and the "Save" button is pressed to activate the draft, the `modifiedAt` field should
be modified.

This behavior remains consistent with both the old draft and the new lean draft implementations when utilizing the
default database layer.

This pull request aims to align the current implementation with the expected behavior outlined above.