cqn4sql: attach adjacent fk adjacent to assoc as element of column
for a flat model, we must attach the foreign key which is adjacent to
it's association as the columns `element` property. If such a flat fk
is found next to the association in the containing entity, it is used
as the `element` rather than the key in the target of the association.

fixes #165