cqn4sql: make sure to always use calculated alias in custom on conditions (#173)
Inspired by #169

---

the issue here was that we had a where condition like

`where exists foo[id=1] or exists foo[id=2]` with foo being an association `foo : Association to one Foo on foo.ID = foo_ID;`.

While building up the where exists subqueries, we calculate unique table aliases for `foo`, which results in a table alias `foo2` for the second condition of the initial where clause. Now, if we incorporate the on-condition into the where clause of the second where exists subquery, we must replace the table alias `foo` from the on-condition with `foo2`.