cqn4sql: skip assocs from wildcard expansion instead of fks in flat mode (#176)
* cqn4sql: skip assocs from wildcard expansion instead of fks in flat mode

this reverts the cqn4sql part of the hotfix in #154

---

The problem was, that in a flat model, the foreign keys were not part of the result anymore, in case that the respective association was overwritten by the smart wildcard mechanism. E.g. by an expand.

The reason was, that in case of a flat model, we skipped the foreign keys and still flattened the association. This caused the bug where overwritting an association caused the lack of the foreign key
in the result.

The fix here is simple, skip the association from the wildcard expansion instead. The compiler already, correctly generated the foreign key elements for us.

* adapt test description

* no debug

* Update db-service/test/cqn4sql/wildcards.test.js

Co-authored-by: Steffen Weinstock <79531202+stewsk@users.noreply.github.com>

---------

Co-authored-by: Steffen Weinstock <79531202+stewsk@users.noreply.github.com>