feat(`UPDATE`): enable path expressions
this change enables the usage of path expressions in
`where` clauses of `UPDATE` queries.

If there is a path expressions found in a `where` of an
`UPDATE`, the `where` clause is cut out of the `UPDATE`
and is put into a `where` clause  in a dummy subselect,
which is then transformed by `cqn4sql`. The result of the
transformation is then put into the `UPDATE` again, as
`exists <subquery>`.
--> `UPDATE.where = ['exists', cqn4sql(subquery)]`

---

for #159