feat(`DELETE`): enable path expressions
this change enables the usage of path expressions in `where` clauses of `DELETE` queries.

If there is a path expressions found in a `where` of an `DELETE`, the `where` clause is cut out of the `DELETE` and is put into a `where` clause  in a dummy subselect, which is then transformed by `cqn4sql`. The result of the transformation is then put into the `DELETE` again, as `exists <subquery>`.
--> `DELETE.where = ['exists', cqn4sql(subquery)]`

---

fix #159