chore: improve error message for struct comparison
the error message calculation was over-enginereed. With little value.
It is sufficient to tell the user that the structures are not
comparable, no need to mention every step which was not found in the msg.
also: exit early once one step was not found in the other side of the
comparison