chore: improve error message for struct comparison (#430)
the error message calculation was over-enginereed. With little value. It
is sufficient to tell the user that the structures are not comparable,
no need to mention every step which was not found in the msg. also: exit
early once one step was not found in the other side of the comparison

**Note:** Best to review this change with whitespaces hidden.