fix: always generate unique subquery aliases (#435)
if we have a query hierarchy, make sure that subqueries always have
aliases which are unique to the outer scope. This does not yield more
functionality for a user, since if we would want to address elements of
an outer query (i.e. make use of correlated subqueries), one has to
define an alias beforehand.

---------

Co-authored-by: Johannes Vogel <31311694+johannes-vogel@users.noreply.github.com>