fix(`cqn4sql`): only transform list if necessary (#438)
because of odata special magic, a tokenstream with `length === 1` which
only has a value, is transformed into a full key tuple comparison, e.g.
to make this possible: `select from bookshop.Books[42]`. If we now have
a list, with exactly one entry, the same logic kicked in and transformed
the single value in the token stream to such a key comparison.

to fix this, we now skip processing of lists which only consist of
values, which circumvents the described issue.