feat: support hybrid access for universal csn
added some tests how one can use universal csn to use structured
and flat access to elements of the underyling entity.

the non-enumerable elements of an entity are marked as "already
unfolded". This marker can be used to skip those elements from
e.g. wildcard expansion but makes it possible to directly access
them in the query.