fix(`cqn4sql`): be robust against `$self.<element>` references
if a `ref` of length > 1, has `ref[0] === 'self'`, the `$self`
can just be ignored. `$self` comparisons have only special semantics
if compared to an association. With this change, we become more robust
in that regard and do not prepend wrong table aliases for the case
`$self.<assoc>.<element>` - in that case, we must ensure that the alias
of the association is used and not the alias of the entity.