fix(`search`): search on aggregated results in `HAVING` clause
If the query contains a `group by`, the search is performed on the queries
columns. If one of those columns is an aggregate function,
we must put the search term in the `HAVING` clause instead of `WHERE`.
Since the logical processing order of SQL queries is `FROM -> WHERE -> GROUP BY -> HAVING -> SELECT -> ORDER BY` the aggregated results are not accessible in the `WHERE` clause.

This led to an error:

```sql
> Uncaught SqliteError: misuse of aggregate: COUNT() in:
SELECT
            COUNT(Books.stock) AS foo
        FROM
            sap_capire_bookshop_Books AS Books
        WHERE
            (IFNULL(INSTR(LOWER(COUNT(stock)), LOWER(?)), 0))
        GROUP BY
            Books.title
```

With this change, the searchable columns of the query are
checked against the  aggregate functions of the ANSI SQL standard.
If one is found and the query contains a `group by` (which should then
always be the case), the search expression is put into the `HAVING` clause.

--> fixes some cds tests