fix(`hana`): use keyword list from compiler (#525)
today we updated the keywords in the compiler because new hana version
came with some new keywords. Instead of copying the keywords to this
repo, we should rather import the list from the compiler to get rid of
this double maintenance.

I already requested that also keywords for other sql dialects will be
available in a similar fashion.