fix(`order by`): for localized sorting, prepend table alias (#546)
Usually, column references in _simple_ names in order by clauses are
left as is. That means **no table alias** is prepended. In expressions
however, **only source elements** can be reffered to.

That is why in case of localized sorting, we must prepend the table
alias nevertheless, as in the DB specific drivers, a `COLLATE`
expression is wrapped around the string reference. Hence we must replace
the column reference with the underlying source reference.

as reported in #543, postgres is not able to resolve a reference to a
column unambigously if the reference is part of a `COLLATE` expression.