fix(`order by`): do not mistake simple name with infix filter
If the query has `columns` but they solely consist
of `*`, `cqn4sql` wrongly assumed that an association
traversal has to be resolved in the queries elements,
this lead to a confusing error complaining that only foreign key
of assocs can be traversed in infix filters.

Actually, if we have something like this:

```
select from Books {
  author as bubu
} order by bubu.name
```

we must reject the non foreign key traversal of `bubu`,
as we can't build the correct join. Instead, a user must use
the association `Books:author`. This change aligns the behavior
with the cds compiler.