test(`cds 8`): explicitly set `fewerLocalizedViews = false`
the `cds-compiler` will not produce `localized` views for
transitive dependencies of `localized` artifacts.

That means an `Author` will not get a `localized.Authors`
view just because there is an `author` association in the
`localized` entity `Books`.

at some point those tests can be rewritten, but to test
the deep replacement mechanism for `localized` artifacts
it is fine for now to just explicitly set the `fewerLocalizedViews`
option.