fix: enforce db file to be existent before server start (#742)
docs:
https://github.com/WiseLibs/better-sqlite3/blob/master/docs/api.md#new-databasepath-options

`options.fileMustExist`: if the database does not exist, an `Error` will
be thrown instead of creating a new file. This option is ignored for
in-memory, temporary, or readonly database connections (default:
`false`).

With this any CRUD request does not create empty db files. It's
obviously an error by developer but does not pollute the environment.