fix(sqlite): default `[development]` URL to `:memory:` (#926)
```sh
cds init bookshop && cd bookshop
cds env get requires.db
```
currently gives you
```js
{
  impl: '@cap-js/sqlite',
  credentials: { url: 'db.sqlite', database: ':memory:' }, 💣 "url" and "database" don't match
  kind: 'sqlite'
}
```

There are a few ways to solve this but I think `@cap-js/sqlite` should
determine itself that the default URL for `[development]` is `:memory:`,
and not rely on the default defined in `@sap/cds` (see the `_databases`
section in `cds-requires.js`), which it currently does.

This is a non-breaking change, as today the `cds`-provided `url:
':memory:'` wins, so even before we deployed to an in-memory DB by
default. This just fixes a confusing log and prepares for the next major
version, where we can then remove the redundant and outdated
`credentials.database`.