perf: Remove `$$RN$$` and additional sub selects (#929)
When there is no need for `$$RN$$` it is excluded from the query. When
no additional internal columns are created the additional sub select is
collapsed into the primary select. Reducing the complexity of all
overall queries. This change mostly applies to queries using
`sql_simple_queries` as they don't require `$$RN$$` for `orderBy`
clauses.

---------

Co-authored-by: Johannes Vogel <31311694+johannes-vogel@users.noreply.github.com>