fix: groupby path expression with overlapping identifier (#992)
Currently, the HANA service is unable to `GROUP BY foobar.foobar`,
raising an error like `invalid column name: foobar.foobar`, if the query
includes an aggregation and `sql syntax error: incorrect syntax near
"FROM"`, if it does not. The problem is caused by the overlapping column
names and the HANA service not correctly injecting the column groupby
requires into the parent select - if the parent includes a column with
the same `column_name` as the 'grouped-by' column.

This fix enables the required injection, by skipping the overlapping
names check for columns that contain a select.

---------

Co-authored-by: Bob den Os <108393871+BobdenOs@users.noreply.github.com>
Co-authored-by: Johannes Vogel <31311694+johannes-vogel@users.noreply.github.com>