feat: pass through of  arbitrary client options (#1024)
Pass through of arbitrary client options should be possible. The
database services do not interpret this but simply pass through
information to the respective db clients.

```json
{
  "db": {
     "kind": "hana",
     "credentials": { ... },
     "client": { "packetSize": 1048576},
  }
}
```
This sample refers to the [`packetSize` parameter of
`hdb`](https://github.com/SAP/node-hdb?tab=readme-ov-file#controlling-the-maximum-packet-size).

For Hana MT and SQLite the "ugly reuse" of credentials is not possible.
In Hana MT, the credentials for the db connection are not in environment
but provided by service manager.
In SQLite, the credentials and options are different parameters for the
client.

SQLite options:
https://github.com/WiseLibs/better-sqlite3/blob/master/docs/api.md#new-databasepath-options
Postgres options: https://node-postgres.com/apis/client#new-client
hdb options:
https://github.com/SAP/node-hdb?tab=readme-ov-file#cesu-8-encoding-support
hana-client options:
https://help.sap.com/docs/PRODUCT_ID/f1b440ded6144a54ada97ff95dac7adf/4fe9978ebac44f35b9369ef5a4a26f4c.html?state=PRODUCTION&version=latest&locale=en-US

closes https://github.com/cap-js/cds-dbs/issues/994
closes https://github.com/cap-js/cds-dbs/issues/920