fix(`expand`): proper subquery `from` construction (#1126)
if a query selects from another query, we need to consider the case
where the subquery has path expressions (and therefore joins).

The best solution is then to drill down into the innermost query of the
from clause, because we do not know the depth.

It is also not a good idea to check for the `ref` of the innermost
_untransformed_ query, as it is potentially a scoped query (as shown in
one of the tests).

Moreover, it is not possible to use the `ref` of the _transformed_ query
because it might be replaced by join `args`.

replaces #1114

fix #1112