fix(`search`): do not search on non-projected elements (#1198)
the `@cds.search` annotation is propagated as-is.

If a path expression was used in the `@cds.search` annotation of a base
entity, but the association for the path is _not_ projected in a view on
top of the entity, we must not try to resolve the path expression in the
view.

reported in cap/issue#18465