chore(cqn4sql): no special handling for `expand` with `groupBy`
to satisfy OData expecations for queries which include `$apply`
in combination with an aggregation + `groupBy`, we added some special
handling to the cqn transformation for all queries which come with
an `expand` column and `groupBy`. This leads to unexpected results,
especially if one does not interact with the DB through the OData layer.

Rather than making assumptions about the intention of the query, `cqn4sql`
should stay protocol agnostic. It is the job of the OData adapter, to parse
such OData queries to a `CQN` in a way, that `cqn4sql` can do its job as-usual.