fix: Allow raw streams to have empty results (#1224)
Currently when a raw Postgres stream doesn't have any rows. It results
in `null` being returned as no `resolve` was called with the `stream`.
This change fixes this behavior by checking whether it has written
anything to the stream yet if not it will put the `[` character onto the
stream to initiate the array directly followed by the `]` character to
finish the array.

Co-authored-by: Johannes Vogel <31311694+johannes-vogel@users.noreply.github.com>