fix: ensure ordering of `ParameterStream` chunks (#1280)
Under certain conditions, the chunks written by the `ParameterStream`
class were transmitted in the wrong order:


<img width="1687" height="1415" alt="image"
src="https://github.com/user-attachments/assets/f049a2a2-1516-474e-ab1f-df20819105dc"
/>
In this example, there was a correct copydata header with length 0x158
written to the stream, but the header of another chunk with length 0x11
was transmitted first.


This leads to errors like these because copydata message
headers/announced lengths do not match the actual message length:
`Error in ParameterStream: error: unexpected message type 0x6B during
COPY from stdin`

I could only observe the issue with TLS enabled (local/BTP) Postgres, so
it looks like TLS sockets are a bit more sensitive to timing. It happens
predictably for me there, though. Also it never happened with the
patches included in this PR, so there is hope that it actually resolves
the issue.

It could be that the code in this PR is a bit too strict in ensuring
sequence, but at least it works for me. I'm not sure how to unit test
this, because the requirements are quite high (I think adding a
TLS-enabled postgres instance to the test environment is asking a bit
too much?)