fix(`assoc2join`): target side access detection (#1282)
There were some cases were the `assoc 2 join` translation wrongly found
a complex path to be a foreign key. That was the case if the FK
semantics were invalidated by a previous filter / paramaeter traversion.

With this change, some systematic tests have been added to make sure
that proper join nodes are generated and the key is properly fetched
from the target side.