fix: set proper element link for path into fk (#1344)
→ in the changed `if` branch, we (re) set the leaf link again to the
last association. However, the path in the column already drilled down
to the respective element. We just need to propagate the already
calculated element and attach it to the flat column.

Before this change, in this particular case the association (`parent` in
the test) was erroneously set as `element` link for the column, which
led to a faulty `cqn2sql` rendering down the line.