feat(`a2j`): enable path expression inside filter
for join relevant paths within infix filters __always__ build
a correlated subquery with the current join node and put the complete
filter condition into the subqueries `where`.

I tested quite some examples with this approach and they all seem to do the trick.
This approach also works for trivial conditions like `where books[author.ID = 5]` for
which we would not necessarily need to do this.