feat(`flattening`): allow flattening of `n`-ary structures in `list` (#1337)
a structured element which expands to `>=1` sub elements / foreign keys
can be expanded to all its leafs in a `list` or in a `group by`.

**Restrictions**:
- in other list like constructs like `function.args` & `order by` this
is only possible for structures with exactly one scalar leaf due to
unclear semantics.

The `getTransformedTokenStream` function has access to the context in
which the token stream shall be expanded. With this information it is
still possible reject a structure/assoc as a value in `having` & `where`

other changes:

feat(`flattening`): allow flattening of structures with exactly one leaf
in `order by` and `args`