fix: fixed versions in postgres deployer package.json (#1429)
We were surprised by fields like `CreatedByUserDescription`,
`LastChangedByUserDescription`, `InProcessByUserDescription` in the
`DraftAdministrativeData` in our deployed DB.

The cause was that without fixed versions, the deployer took the most
recent `@sap/cds` (from the "^9" range, which is 9.3.x right now). 9.3
added the fields mentioned above (see Capire changelog), while we still
are (and want to be) on 9.2.x.

This is one potential solution (use the same versions as we had them
during build), but I'm open for different ones!

---------

Co-authored-by: Steffen Waldmann <steffenwaldmann@me.com>
Co-authored-by: Patrice Bender <info@patrice.codes>