fix(cqn4sql): calculated elements with function expr in from
forgot to propagate the context ("resolve calculated element in from")
that caused the resolver to lookup the first `ref` step in the queries
sources (table aliases) - which is generally not possible in a from clause.

TODO: last step of `from.ref` may actually contain a table alias reference,
as this is just syntactic sugar for a separate `where`. I will tackle this
in another change, though.